# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Together API",
    "description": "Private, non-federated Discord alternative for small gaming communities",
    "contact": {
      "name": "Together Contributors"
    },
    "license": {
      "name": "PolyForm Noncommercial 1.0.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/channels/{channel_id}/events": {
      "post": {
        "tags": [
          "Events"
        ],
        "operationId": "create_event",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/servers/{id}/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "list_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of upcoming events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerEventDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEventPayload": {
        "type": "object",
        "required": [
          "name",
          "starts_at"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MessageDto": {
        "type": "object",
        "description": "API response for a message. Wraps Message with optional rich content.",
        "required": [
          "id",
          "channel_id",
          "content",
          "mention_user_ids",
          "mention_everyone",
          "thread_reply_count",
          "deleted",
          "created_at",
          "pinned"
        ],
        "properties": {
          "author_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "channel_id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted": {
            "type": "boolean"
          },
          "edited_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "event": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ServerEventDto",
                "description": "Some when the message was created by /event"
              }
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "mention_everyone": {
            "type": "boolean"
          },
          "mention_user_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "pinned": {
            "type": "boolean"
          },
          "pinned_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "pinned_by": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "poll": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PollDto",
                "description": "Some when the message was created by /poll"
              }
            ]
          },
          "reply_to": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "thread_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "thread_reply_count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PollDto": {
        "type": "object",
        "required": [
          "id",
          "question",
          "options",
          "total_votes"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PollOptionDto"
            }
          },
          "question": {
            "type": "string"
          },
          "total_votes": {
            "type": "integer",
            "format": "int64"
          },
          "user_vote": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The option_id the calling user voted for, or None"
          }
        }
      },
...
      "PollOptionDto": {
        "type": "object",
        "required": [
          "id",
          "text",
          "votes"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string"
          },
          "votes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ServerEventDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "starts_at",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}